moved sql from home to timetables view
