+++ home/urls.py
@@ -1,4 +1,4 @@
-## APP (home)
+# home urls.py
 
 from django.urls import path
 from . import views
