+++ timetables/urls.py
@@ -1,0 +1,8 @@
+# timetables urls.py
+
+from django.urls import path
+from . import views
+
+urlpatterns = [
+    path('', views.timetables, name='timetables'),
+]
